[refactor] 신고 관련 리팩토링
