+++ src/main/java/org/wooriverygood/api/report/domain/PostReport.java
@@ -1,6 +1,7 @@
 package org.wooriverygood.api.report.domain;
 
 import jakarta.persistence.*;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,7 +9,7 @@ import org.wooriverygood.api.post.domain.Post;
 
 @Entity
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PostReport {
 
     @Id
